+++ src/TDBMService.php
@@ -546,6 +546,12 @@ class TDBMService
         if (isset($this->tableToBeanMap[$tableName])) {
             return $this->tableToBeanMap[$tableName];
         } else {
+            $key = $this->cachePrefix.'_tableToBean_'.$tableName;
+            $cache = $this->cache->fetch($key);
+            if ($cache) {
+                return $cache;
+            }
+
             $className = $this->beanNamespace.'\\'.$this->namingStrategy->getBeanClassName($tableName);
 
             if (!class_exists($className)) {
@@ -553,6 +559,7 @@ class TDBMService
             }
 
             $this->tableToBeanMap[$tableName] = $className;
+            $this->cache->save($key, $className);
             return $className;
         }
     }
@@ -1421,19 +1428,6 @@ class TDBMService
         }
 
         return $item;
-    }
-
-    /**
-     * Returns the foreign key object.
-     *
-     * @param string $table
-     * @param string $fkName
-     *
-     * @return ForeignKeyConstraint
-     */
-    public function _getForeignKeyByName(string $table, string $fkName): ForeignKeyConstraint
-    {
-        return $this->tdbmSchemaAnalyzer->getSchema()->getTable($table)->getForeignKey($fkName);
     }
 
     /**