+++ src/Utils/Annotation/JsonIgnore.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TheCodingMachine\TDBM\Utils\Annotation;
+
+/**
+ * @Annotation
+ */
+final class JsonIgnore
+{
+}