+++ src/Utils/BeanDescriptor.php
@@ -17,6 +17,8 @@ use TheCodingMachine\TDBM\AlterableResultIterator;
 use TheCodingMachine\TDBM\ConfigurationInterface;
 use TheCodingMachine\TDBM\ResultIterator;
 use TheCodingMachine\TDBM\SafeFunctions;
+use TheCodingMachine\TDBM\Schema\ForeignKey;
+use TheCodingMachine\TDBM\Schema\ForeignKeys;
 use TheCodingMachine\TDBM\TDBMException;
 use TheCodingMachine\TDBM\TDBMSchemaAnalyzer;
 use TheCodingMachine\TDBM\TDBMService;
@@ -26,6 +28,7 @@ use TheCodingMachine\TDBM\Utils\Annotation\AddTrait;
 use TheCodingMachine\TDBM\Utils\Annotation\AddTraitOnDao;
 use TheCodingMachine\TDBM\Utils\Annotation\AnnotationParser;
 use TheCodingMachine\TDBM\Utils\Annotation\AddInterface;
+use Zend\Code\Generator\AbstractMemberGenerator;
 use Zend\Code\Generator\ClassGenerator;
 use Zend\Code\Generator\DocBlock\Tag\ParamTag;
 use Zend\Code\Generator\DocBlock\Tag\ReturnTag;
@@ -377,7 +380,7 @@ class BeanDescriptor implements BeanDescriptorInterface
                 continue;
             }
 
-            $descs[] = new PivotTableMethodsDescriptor($table, $localFk, $remoteFk, $this->namingStrategy, $this->beanNamespace);
+            $descs[] = new PivotTableMethodsDescriptor($table, $localFk, $remoteFk, $this->namingStrategy, $this->beanNamespace, $this->annotationParser);
         }
 
         return $descs;
@@ -417,11 +420,6 @@ class BeanDescriptor implements BeanDescriptorInterface
     {
         $tableName = $this->table->getName();
         $parentFk = $this->schemaAnalyzer->getParentRelationship($tableName);
-        if ($parentFk !== null) {
-            $initializer = '$array = parent::jsonSerialize($stopRecursion);';
-        } else {
-            $initializer = '$array = [];';
-        }
 
         $method = new MethodGenerator('jsonSerialize');
         $method->setDocBlock('Serializes the object for JSON encoding.');
@@ -429,24 +427,30 @@ class BeanDescriptor implements BeanDescriptorInterface
         $method->getDocBlock()->setTag(new ReturnTag(['array']));
         $method->setParameter(new ParameterGenerator('stopRecursion', 'bool', false));
 
-        $str = '%s
-%s
-%s
-return $array;
-';
+        if ($parentFk !== null) {
+            $body = '$array = parent::jsonSerialize($stopRecursion);';
+        } else {
+            $body = '$array = [];';
+        }
 
-        $propertiesCode = '';
         foreach ($this->getExposedProperties() as $beanPropertyDescriptor) {
-            $propertiesCode .= $beanPropertyDescriptor->getJsonSerializeCode();
+            $propertyCode = $beanPropertyDescriptor->getJsonSerializeCode();
+            if (!empty($propertyCode)) {
+                $body .= PHP_EOL . $propertyCode;
+            }
         }
 
         // Many2many relationships
-        $methodsCode = '';
         foreach ($this->getMethodDescriptors() as $methodDescriptor) {
-            $methodsCode .= $methodDescriptor->getJsonSerializeCode();
+            $methodCode = $methodDescriptor->getJsonSerializeCode();
+            if (!empty($methodCode)) {
+                $body .= PHP_EOL . $methodCode;
+            }
         }
 
-        $method->setBody(sprintf($str, $initializer, $propertiesCode, $methodsCode));
+        $body .= PHP_EOL . 'return $array;';
+
+        $method->setBody($body);
 
         return $method;
     }
@@ -522,6 +526,7 @@ return $array;
         $file->setUse(AlterableResultIterator::class);
         $file->setUse(Uuid::class);
         $file->setUse(JsonSerializable::class);
+        $file->setUse(ForeignKeys::class);
 
         $class->setName($baseClassName);
         $class->setAbstract(true);
@@ -552,7 +557,11 @@ EOF
             $class->addMethodFromGenerator($this->generateBeanConstructor());
         }
 
+        $fks = [];
         foreach ($this->getExposedProperties() as $property) {
+            if ($property instanceof ObjectBeanPropertyDescriptor) {
+                $fks[] = $property->getForeignKey();
+            }
             [$getter, $setter] = $property->getGetterSetterCode();
             [$getter, $setter] = $this->codeGeneratorListener->onBaseBeanPropertyGenerated($getter, $setter, $property, $this, $this->configuration, $class);
             if ($getter !== null) {
@@ -582,6 +591,15 @@ EOF
                 throw new \RuntimeException('Unexpected instance'); // @codeCoverageIgnore
             }
         }
+
+        $foreignKeysProperty = new PropertyGenerator('foreignKeys');
+        $foreignKeysProperty->setStatic(true);
+        $foreignKeysProperty->setVisibility(AbstractMemberGenerator::VISIBILITY_PRIVATE);
+        $foreignKeysProperty->setDocBlock(new DocBlockGenerator(null, null, [new VarTag(null, ['\\'.ForeignKeys::class])]));
+        $class->addPropertyFromGenerator($foreignKeysProperty);
+
+        $method = $this->generateGetForeignKeys($fks);
+        $class->addMethodFromGenerator($method);
 
         $method = $this->generateJsonSerialize();
         $method = $this->codeGeneratorListener->onBaseBeanJsonSerializeGenerated($method, $this, $this->configuration, $class);
@@ -1256,7 +1274,8 @@ return $tables;', var_export($this->table->getName(), true));
         $relationships = $this->getPropertiesForTable($this->table);
         foreach ($relationships as $relationship) {
             if ($relationship instanceof ObjectBeanPropertyDescriptor) {
-                $code .= sprintf('$this->setRef('.var_export($relationship->getForeignKey()->getName(), true).', null, '.var_export($this->table->getName(), true).");\n");
+                $tdbmFk = ForeignKey::createFromFk($relationship->getForeignKey());
+                $code .= '$this->setRef('.var_export($tdbmFk->getCacheKey(), true).', null, '.var_export($this->table->getName(), true).");\n";
             }
         }
 
@@ -1367,5 +1386,73 @@ return $tables;', var_export($this->table->getName(), true));
     public function getGeneratedBeanNamespace(): string
     {
         return $this->generatedBeanNamespace;
+    }
+
+    /**
+     * @param ForeignKeyConstraint[] $fks
+     */
+    private function generateGetForeignKeys(array $fks): MethodGenerator
+    {
+        $fkArray = [];
+
+        foreach ($fks as $fk) {
+            $tdbmFk = ForeignKey::createFromFk($fk);
+            $fkArray[$tdbmFk->getCacheKey()] = [
+                ForeignKey::FOREIGN_TABLE => $fk->getForeignTableName(),
+                ForeignKey::LOCAL_COLUMNS => $fk->getUnquotedLocalColumns(),
+                ForeignKey::FOREIGN_COLUMNS => $fk->getUnquotedForeignColumns(),
+            ];
+        }
+
+        ksort($fkArray);
+        foreach ($fkArray as $tableFks) {
+            ksort($tableFks);
+        }
+
+        $code = <<<EOF
+if (\$tableName === %s) {
+    if (self::\$foreignKeys === null) {
+        self::\$foreignKeys = new ForeignKeys(%s);
+    }
+    return self::\$foreignKeys;
+}
+return parent::getForeignKeys(\$tableName);
+EOF;
+        $code = sprintf($code, var_export($this->getTable()->getName(), true), $this->psr2VarExport($fkArray, '    '));
+
+        $method = new MethodGenerator('getForeignKeys');
+        $method->setVisibility(AbstractMemberGenerator::VISIBILITY_PROTECTED);
+        $method->setStatic(true);
+        $method->setDocBlock('Internal method used to retrieve the list of foreign keys attached to this bean.');
+        $method->setReturnType(ForeignKeys::class);
+
+        $parameter = new ParameterGenerator('tableName');
+        $parameter->setType('string');
+        $method->setParameter($parameter);
+
+
+        $method->setBody($code);
+
+        return $method;
+    }
+
+    /**
+     * @param mixed $var
+     * @param string $indent
+     * @return string
+     */
+    private function psr2VarExport($var, string $indent=''): string
+    {
+        if (is_array($var)) {
+            $indexed = array_keys($var) === range(0, count($var) - 1);
+            $r = [];
+            foreach ($var as $key => $value) {
+                $r[] = "$indent    "
+                    . ($indexed ? '' : $this->psr2VarExport($key) . ' => ')
+                    . $this->psr2VarExport($value, "$indent    ");
+            }
+            return "[\n" . implode(",\n", $r) . "\n" . $indent . ']';
+        }
+        return var_export($var, true);
     }
 }