+++ src/Utils/DirectForeignKeyMethodDescriptor.php
@@ -7,8 +7,10 @@ namespace TheCodingMachine\TDBM\Utils;
 use Doctrine\DBAL\Schema\ForeignKeyConstraint;
 use Doctrine\DBAL\Schema\Table;
 use TheCodingMachine\TDBM\AlterableResultIterator;
+use TheCodingMachine\TDBM\Schema\ForeignKey;
 use TheCodingMachine\TDBM\TDBMException;
 use TheCodingMachine\TDBM\Utils\Annotation\AnnotationParser;
+use TheCodingMachine\TDBM\Utils\Annotation;
 use Zend\Code\Generator\AbstractMemberGenerator;
 use Zend\Code\Generator\DocBlock\Tag\ReturnTag;
 use Zend\Code\Generator\MethodGenerator;
@@ -107,11 +109,12 @@ class DirectForeignKeyMethodDescriptor implements MethodDescriptorInterface
         ]));
         $getter->setReturnType(AlterableResultIterator::class);
 
+        $tdbmFk = ForeignKey::createFromFk($this->foreignKey);
+
         $code = sprintf(
-            'return $this->retrieveManyToOneRelationshipsStorage(%s, %s, %s, %s);',
+            'return $this->retrieveManyToOneRelationshipsStorage(%s, %s, %s);',
             var_export($this->foreignKey->getLocalTableName(), true),
-            var_export($this->foreignKey->getName(), true),
-            var_export($this->foreignKey->getLocalTableName(), true),
+            var_export($tdbmFk->getCacheKey(), true),
             $this->getFilters($this->foreignKey)
         );
 
@@ -158,7 +161,40 @@ class DirectForeignKeyMethodDescriptor implements MethodDescriptorInterface
      */
     public function getJsonSerializeCode() : string
     {
-        return '';
+        /** @var Annotation\JsonCollection|null $jsonCollection */
+        $jsonCollection = $this->findAnnotation(Annotation\JsonCollection::class);
+        if ($jsonCollection === null) {
+            return '';
+        }
+
+        /** @var Annotation\JsonFormat|null $jsonFormat */
+        $jsonFormat = $this->findAnnotation(Annotation\JsonFormat::class);
+        if ($jsonFormat !== null) {
+            $method = $jsonFormat->method ?? 'get' . ucfirst($jsonFormat->property);
+            $format = "$method()";
+        } else {
+            $stopRecursion = $this->findAnnotation(Annotation\JsonRecursive::class) ? '' : 'true';
+            $format = "jsonSerialize($stopRecursion)";
+        }
+        $isIncluded = $this->findAnnotation(Annotation\JsonInclude::class) !== null;
+        $index = $jsonCollection->key ?: lcfirst(TDBMDaoGenerator::toCamelCase($this->foreignKey->getLocalTableName()));
+        $class = $this->getBeanClassName();
+        $variableName = '$' . TDBMDaoGenerator::toVariableName($class);
+        $getter = $this->getName();
+        $code = <<<PHP
+\$array['$index'] = array_map(function ($class $variableName) {
+    return ${variableName}->$format;
+}, \$this->$getter()->toArray());
+PHP;
+        if (!$isIncluded) {
+            $code = preg_replace('(\n)', '\0    ', $code);
+            $code = <<<PHP
+if (!\$stopRecursion) {
+    $code
+};
+PHP;
+        }
+        return $code;
     }
 
     /**
@@ -180,11 +216,21 @@ class DirectForeignKeyMethodDescriptor implements MethodDescriptorInterface
 
     private function isProtected(): bool
     {
+        return $this->findAnnotation(Annotation\ProtectedOneToMany::class) !== null;
+    }
+
+    /**
+     * @param string $type
+     * @return null|object
+     */
+    private function findAnnotation(string $type)
+    {
         foreach ($this->getAnnotations() as $annotations) {
-            if ($annotations->findAnnotation(Annotation\ProtectedOneToMany::class)) {
-                return true;
+            $annotation = $annotations->findAnnotation($type);
+            if ($annotation !== null) {
+                return $annotation;
             }
         }
-        return false;
+        return null;
     }
 }